Move setup container to separate component
